main 메서드 생성 및 ResultView 코드 분리
